・不要なコードを削除 ・WageViewEntityの作成
