+++ ios/Parjob/Parjob/User/Wage/WageViewEntity.swift
@@ -1,0 +1,18 @@
+//
+//  WageViewEntity.swift
+//  Parjob
+//
+//  Created by 岩見建汰 on 2018/07/02.
+//  Copyright © 2018年 Kenta Iwami. All rights reserved.
+//
+
+import Foundation
+
+struct UserWage {
+    var daytimeStart: String = ""
+    var daytimeEnd: String = ""
+    var nightStart: String = ""
+    var nightEnd: String = ""
+    var daytimeWage: Int = 0
+    var nightWage: Int = 0
+}
